feat: implement VaultDetailContent component and refactor vault detail page
